Patch CombatDirector.Init AND .ResetEliteType
Some mods can cause `CombatDirector.Init()` to be called multiple times,
and each call resets the `eliteTiers` array, thus undoing the change
that this mod introduces. Instead of only modifying the array once, in
`ResetEliteType()`, we instead try to do it in `Init()`. If the content
has not yet loaded, defer initialization until `ResetEliteType()` is
called. It's not ideal, but something something maximum compatibility, I
hope.